refactor embedded metabot components into separate files
