+++ enterprise/frontend/src/embedding-sdk-bundle/components/public/MetabotQuestion/MetabotQuestion.tsx
@@ -1,39 +1,20 @@
 import cx from "classnames";
-import { useEffect, useId, useMemo, useRef } from "react";
-import { t } from "ttag";
+import { useId } from "react";
 
 import { FlexibleSizeComponent } from "embedding-sdk-bundle/components/private/FlexibleSizeComponent";
 import { withPublicComponentWrapper } from "embedding-sdk-bundle/components/private/PublicComponentWrapper";
 import { SdkAdHocQuestion } from "embedding-sdk-bundle/components/private/SdkAdHocQuestion";
 import { SdkQuestionDefaultView } from "embedding-sdk-bundle/components/private/SdkQuestionDefaultView";
-import { useSdkDispatch } from "embedding-sdk-bundle/store";
 import { EnsureSingleInstance } from "embedding-sdk-shared/components/EnsureSingleInstance/EnsureSingleInstance";
 import useIsSmallScreen from "metabase/common/hooks/use-is-small-screen";
 import { useLocale } from "metabase/common/hooks/use-locale";
-import {
-  Button,
-  Flex,
-  Icon,
-  Loader,
-  Menu,
-  Stack,
-  Text,
-  Textarea,
-  Tooltip,
-  UnstyledButton,
-} from "metabase/ui";
-import { useGetSuggestedMetabotPromptsQuery } from "metabase-enterprise/api";
-import { Messages } from "metabase-enterprise/metabot/components/MetabotChat/MetabotChatMessage";
-import { MetabotResetLongChatButton } from "metabase-enterprise/metabot/components/MetabotChat/MetabotResetLongChatButton";
-import {
-  useMetabotAgent,
-  useMetabotChatHandlers,
-} from "metabase-enterprise/metabot/hooks";
+import { Stack } from "metabase/ui";
 import { useMetabotReactions } from "metabase-enterprise/metabot/hooks/use-metabot-reactions";
-import { cancelInflightAgentRequests } from "metabase-enterprise/metabot/state";
 
 import S from "./MetabotQuestion.module.css";
 import { metabotQuestionSchema } from "./MetabotQuestion.schema";
+import { MetabotQuestionEmptyState } from "./MetabotQuestionEmptyState";
+import { MetabotSidebar } from "./MetabotSidebar";
 import { QuestionDetails } from "./QuestionDetails";
 import { QuestionTitle } from "./QuestionTitle";
 import type { MetabotQuestionProps } from "./types";
@@ -102,274 +83,6 @@ const MetabotQuestionInner = ({
   );
 };
 
-function MetabotSidebar() {
-  const metabot = useMetabotAgent();
-  const { handleSubmitInput } = useMetabotChatHandlers();
-
-  const suggestedPromptsQuery = useGetSuggestedMetabotPromptsQuery({
-    metabot_id: metabot.metabotId,
-    limit: 3,
-    sample: true,
-  });
-
-  const suggestedPrompts = useMemo(() => {
-    return suggestedPromptsQuery.currentData?.prompts ?? [];
-  }, [suggestedPromptsQuery]);
-
-  const hasMessages =
-    metabot.messages.length > 0 || metabot.errorMessages.length > 0;
-
-  return (
-    <Stack className={S.sidebar}>
-      <SidebarHeader />
-      <SidebarChatHistory />
-      <SidebarInput
-        suggestedPrompts={suggestedPrompts}
-        hasMessages={hasMessages}
-        onSubmitPrompt={handleSubmitInput}
-      />
-    </Stack>
-  );
-}
-
-function SidebarHeader() {
-  const metabot = useMetabotAgent();
-
-  const startNewConversation = () => {
-    metabot.resetConversation();
-  };
-
-  return (
-    <Flex
-      px="md"
-      py="sm"
-      justify="space-between"
-      align="center"
-      className={S.sidebarHeader}
-    >
-      <Text fz="sm" c="var(--mb-color-text-tertiary)">
-        {t`AI isn't perfect. Double-check results.`}
-      </Text>
-
-      <Tooltip label={t`Start new chat`}>
-        <UnstyledButton
-          onClick={startNewConversation}
-          data-testid="metabot-new-conversation"
-        >
-          <Icon
-            name="edit_document_outlined"
-            size="1rem"
-            c="var(--mb-color-text-tertiary)"
-          />
-        </UnstyledButton>
-      </Tooltip>
-    </Flex>
-  );
-}
-
-function SidebarChatHistory() {
-  const metabot = useMetabotAgent();
-  const { messages, errorMessages } = metabot;
-  const { handleRetryMessage } = useMetabotChatHandlers();
-  const { setNavigateToPath } = useMetabotReactions();
-  const scrollContainerRef = useRef<HTMLDivElement>(null);
-
-  const hasMessages = messages.length > 0 || errorMessages.length > 0;
-
-  // Auto-scroll to bottom when new messages are received
-  useEffect(() => {
-    if (scrollContainerRef.current) {
-      scrollContainerRef.current.scrollTop =
-        scrollContainerRef.current.scrollHeight;
-    }
-  }, [messages.length, errorMessages.length, metabot.isDoingScience]);
-
-  return (
-    <Stack
-      ref={scrollContainerRef}
-      flex={1}
-      gap={0}
-      style={{ overflowY: "auto" }}
-      p="md"
-    >
-      {hasMessages ? (
-        <Messages
-          messages={messages}
-          errorMessages={errorMessages}
-          onRetryMessage={handleRetryMessage}
-          isDoingScience={metabot.isDoingScience}
-          showFeedbackButtons={false}
-          onInternalLinkClick={setNavigateToPath}
-        />
-      ) : null}
-      {metabot.isLongConversation && <MetabotResetLongChatButton />}
-    </Stack>
-  );
-}
-
-function SidebarInput({
-  suggestedPrompts,
-  hasMessages,
-  onSubmitPrompt,
-}: {
-  suggestedPrompts: Array<{ prompt: string }>;
-  hasMessages: boolean;
-  onSubmitPrompt: (prompt: string) => void;
-}) {
-  const metabot = useMetabotAgent();
-  const { handleSubmitInput, handleResetInput } = useMetabotChatHandlers();
-  const dispatch = useSdkDispatch();
-
-  const placeholder = metabot.isDoingScience
-    ? t`Doing science...`
-    : t`Ask AI a question...`;
-
-  const cancelRequest = () => {
-    dispatch(cancelInflightAgentRequests());
-  };
-
-  const resetInput = () => {
-    handleResetInput();
-  };
-
-  const startNewConversation = () => {
-    metabot.resetConversation();
-  };
-
-  return (
-    <Stack gap={0}>
-      {/* Suggested prompts - show only when no messages and not doing science */}
-      {!hasMessages &&
-        !metabot.isDoingScience &&
-        suggestedPrompts.length > 0 && (
-          <Stack gap="sm" p="md" className={S.promptSuggestionsContainer}>
-            {suggestedPrompts.map(({ prompt }, index) => (
-              <Button
-                key={index}
-                size="xs"
-                variant="outline"
-                fw={400}
-                onClick={() => onSubmitPrompt(prompt)}
-                className={S.promptSuggestionButton}
-              >
-                {prompt}
-              </Button>
-            ))}
-          </Stack>
-        )}
-
-      {/* Input area */}
-      <Flex
-        gap="xs"
-        px="md"
-        pt="0.6rem"
-        pb="0.2rem"
-        style={{ borderTop: "1px solid var(--mb-color-border)" }}
-        align="center"
-        justify="center"
-      >
-        <Flex
-          style={{ flexShrink: 0, marginBottom: "8px" }}
-          justify="center"
-          align="center"
-        >
-          {metabot.isDoingScience ? (
-            <Loader size="sm" />
-          ) : (
-            <Icon name="ai" c="var(--mb-color-brand)" size="1rem" />
-          )}
-        </Flex>
-
-        <Textarea
-          id="metabot-chat-input"
-          data-testid="metabot-chat-input"
-          w="100%"
-          autosize
-          minRows={1}
-          maxRows={4}
-          ref={metabot.promptInputRef}
-          autoFocus
-          value={metabot.prompt}
-          disabled={metabot.isDoingScience}
-          placeholder={placeholder}
-          onChange={(e) => metabot.setPrompt(e.target.value)}
-          onKeyDown={(e) => {
-            if (e.nativeEvent.isComposing) {
-              return;
-            }
-
-            if (e.key === "Enter") {
-              e.preventDefault();
-              e.stopPropagation();
-              handleSubmitInput(metabot.prompt);
-            }
-          }}
-          styles={{
-            input: {
-              border: "none",
-              borderRadius: "0",
-              backgroundColor: "transparent",
-              "&:focus": {
-                outline: "none",
-                borderColor: "transparent",
-              },
-            },
-          }}
-        />
-
-        {metabot.isDoingScience && (
-          <UnstyledButton
-            h="1rem"
-            data-testid="metabot-cancel-request"
-            onClick={cancelRequest}
-            style={{ marginBottom: "8px" }}
-          >
-            <Tooltip label={t`Stop generation`}>
-              <Icon
-                name="stop"
-                c="var(--mb-color-text-secondary)"
-                size="1rem"
-              />
-            </Tooltip>
-          </UnstyledButton>
-        )}
-
-        {!metabot.isDoingScience && metabot.prompt.length > 0 && (
-          <UnstyledButton
-            h="1rem"
-            onClick={resetInput}
-            data-testid="metabot-close-chat"
-            style={{ marginBottom: "8px" }}
-          >
-            <Icon name="close" c="var(--mb-color-text-secondary)" size="1rem" />
-          </UnstyledButton>
-        )}
-
-        {/* Overflow menu - only shown in stacked layout and auto layout mobile */}
-        <Menu position="top-end" withinPortal>
-          <Menu.Target>
-            <UnstyledButton
-              data-testid="metabot-overflow-button"
-              className={S.stackedOverflowButton}
-            >
-              <Icon name="ellipsis" c="var(--mb-color-text-secondary)" />
-            </UnstyledButton>
-          </Menu.Target>
-
-          <Menu.Dropdown>
-            <Menu.Item
-              leftSection={<Icon name="edit_document_outlined" size="1rem" />}
-              onClick={startNewConversation}
-            >
-              {t`Start new chat`}
-            </Menu.Item>
-          </Menu.Dropdown>
-        </Menu>
-      </Flex>
-    </Stack>
-  );
-}
-
 const MetabotQuestionWrapped = (props: MetabotQuestionProps) => {
   const ensureSingleInstanceId = useId();
 
@@ -389,15 +102,4 @@ const MetabotQuestionWrapped = (props: MetabotQuestionProps) => {
 export const MetabotQuestion = Object.assign(
   withPublicComponentWrapper(MetabotQuestionWrapped),
   { schema: metabotQuestionSchema },
-);
-
-export const MetabotQuestionEmptyState = () => (
-  <Stack h="100%" w="100%" gap="lg" align="center" justify="center">
-    <Icon name="ai" c="var(--mb-color-bg-black)" size="5rem" opacity={0.25} />
-
-    <Stack gap="xs" align="center">
-      <Text lh="md">{t`Ask questions to AI.`}</Text>
-      <Text lh="md">{t`Results will appear here.`}</Text>
-    </Stack>
-  </Stack>
 );
